test(e2e): split live trace workflow guard tests
