+++ test/e2e/support/e2e-workflow.test.ts
@@ -1037,101 +1037,6 @@ jobs:
     }
   });
 
-  it("rejects live trace setup after workspace preparation", () => {
-    const tmp = fs.mkdtempSync(path.join(os.tmpdir(), "e2e-workflow-"));
-    const workflowPath = path.join(tmp, "workflow.yaml");
-    const workflow = readWorkflow() as {
-      jobs: Record<string, { steps: Array<Record<string, unknown>> }>;
-    };
-    const steps = workflow.jobs.live?.steps;
-    expect(steps).toEqual(expect.any(Array));
-    const configureIndex = steps.findIndex(
-      (step) => step.name === "Configure live E2E trace directory",
-    );
-    expect(configureIndex).toBeGreaterThanOrEqual(0);
-    const [configureStep] = steps.splice(configureIndex, 1);
-    const prepareIndex = steps.findIndex((step) => step.name === "Prepare E2E workspace");
-    expect(prepareIndex).toBeGreaterThanOrEqual(0);
-    steps.splice(prepareIndex + 1, 0, configureStep);
-    fs.writeFileSync(workflowPath, YAML.stringify(workflow));
-
-    try {
-      expect(validateE2eWorkflowBoundary(workflowPath)).toContain(
-        "live trace setup, workspace preparation, Vitest run, sanitizer, and cleanup steps must stay in order",
-      );
-    } finally {
-      fs.rmSync(tmp, { recursive: true, force: true });
-    }
-  });
-
-  it("rejects live trace sanitizer without the workflow-owned source guard", () => {
-    const tmp = fs.mkdtempSync(path.join(os.tmpdir(), "e2e-workflow-"));
-    const workflowPath = path.join(tmp, "workflow.yaml");
-    const workflow = readWorkflow() as {
-      jobs: Record<string, { steps: Array<Record<string, unknown>> }>;
-    };
-    const sanitizeStep = workflow.jobs.live?.steps.find(
-      (step) => step.name === "Build trusted live E2E timing summary",
-    );
-    expect(sanitizeStep?.run).toEqual(expect.any(String));
-    sanitizeStep!.run = String(sanitizeStep!.run)
-      .replace('expected_trace_dir="${RUNNER_TEMP}/nemoclaw-e2e-traces/${TARGET_ID}"\n', "")
-      .replace(
-        'if [ -z "${RUNNER_TEMP}" ] || [ "${NEMOCLAW_TRACE_DIR}" != "${expected_trace_dir}" ]; then\n' +
-          '  echo "::error::Refusing to sanitize unexpected raw trace path" >&2\n' +
-          "  exit 1\n" +
-          "fi\n",
-        "",
-      );
-    fs.writeFileSync(workflowPath, YAML.stringify(workflow));
-
-    try {
-      expect(validateE2eWorkflowBoundary(workflowPath)).toEqual(
-        expect.arrayContaining([
-          "step 'Build trusted live E2E timing summary' run script must include ${RUNNER_TEMP}/nemoclaw-e2e-traces/${TARGET_ID}",
-          'step \'Build trusted live E2E timing summary\' run script must include [ "${NEMOCLAW_TRACE_DIR}" != "${expected_trace_dir}" ]',
-        ]),
-      );
-    } finally {
-      fs.rmSync(tmp, { recursive: true, force: true });
-    }
-  });
-
-  it("rejects live trace sanitizer when the source guard moves after Python reads traces", () => {
-    const tmp = fs.mkdtempSync(path.join(os.tmpdir(), "e2e-workflow-"));
-    const workflowPath = path.join(tmp, "workflow.yaml");
-    const workflow = readWorkflow() as {
-      jobs: Record<string, { steps: Array<Record<string, unknown>> }>;
-    };
-    const sanitizeStep = workflow.jobs.live?.steps.find(
-      (step) => step.name === "Build trusted live E2E timing summary",
-    );
-    expect(sanitizeStep?.run).toEqual(expect.any(String));
-    const guardBlock =
-      'expected_trace_dir="${RUNNER_TEMP}/nemoclaw-e2e-traces/${TARGET_ID}"\n' +
-      'if [ -z "${RUNNER_TEMP}" ] || [ "${NEMOCLAW_TRACE_DIR}" != "${expected_trace_dir}" ]; then\n' +
-      '  echo "::error::Refusing to sanitize unexpected raw trace path" >&2\n' +
-      "  exit 1\n" +
-      "fi\n";
-    sanitizeStep!.run = String(sanitizeStep!.run).replace(guardBlock, "") + guardBlock;
-    fs.writeFileSync(workflowPath, YAML.stringify(workflow));
-
-    try {
-      expect(validateE2eWorkflowBoundary(workflowPath)).toEqual(
-        expect.arrayContaining([
-          "step 'Build trusted live E2E timing summary' run script must include " +
-            'expected_trace_dir="${RUNNER_TEMP}/nemoclaw-e2e-traces/${TARGET_ID}" before ' +
-            "python3 scripts/e2e/sanitize-trace-timing.py",
-          "step 'Build trusted live E2E timing summary' run script must include " +
-            '[ "${NEMOCLAW_TRACE_DIR}" != "${expected_trace_dir}" ] before ' +
-            "python3 scripts/e2e/sanitize-trace-timing.py",
-        ]),
-      );
-    } finally {
-      fs.rmSync(tmp, { recursive: true, force: true });
-    }
-  });
-
   it("requires snapshot commands workflow boundary coverage", () => {
     const tmp = fs.mkdtempSync(path.join(os.tmpdir(), "e2e-workflow-"));
     const workflowPath = path.join(tmp, "workflow.yaml");
